OPHAKTKEH-182: review fixes
